Use zustand library to totally make vehicles component dumb
