+++ apps/webapp/src/features/vehicles/di.ts
@@ -8,7 +8,8 @@ import {
   ToggleBreakInMonitoringUseCase,
 } from './domain/use-cases/vehicles.use-cases';
 
-// Provide a singleton instance
+import { createVehiclesStore } from './presentation/store/vehicles.store';
+
 export const vehicleRepository = new VehicleApiRepository(apiClient);
 
 export const getVehiclesUseCase = new GetVehiclesUseCase(vehicleRepository);
@@ -16,3 +17,10 @@ export const configureTelemetryUseCase = new ConfigureTelemetryUseCase(vehicleRe
 export const checkTelemetryConfigUseCase = new CheckTelemetryConfigUseCase(vehicleRepository);
 export const deleteTelemetryConfigUseCase = new DeleteTelemetryConfigUseCase(vehicleRepository);
 export const toggleBreakInMonitoringUseCase = new ToggleBreakInMonitoringUseCase(vehicleRepository);
+
+export const useVehiclesStore = createVehiclesStore({
+  getVehiclesUseCase,
+  configureTelemetryUseCase,
+  deleteTelemetryConfigUseCase,
+  toggleBreakInMonitoringUseCase,
+});
